requirements.txt
plotly==5.10.0
Pillow==9.3.0
numpy==1.25.0
pandas==2.0
matplotlib==3.7.1